feat: telegram bridge

+++ cmd/smpp-telegram-bridge/types.go
@@ -2,10 +2,10 @@ package main
 
 type Configuration struct {
 	Token          string  `json:"token"`
-	AllowedUserIDs []int64 `json:"allowed_user_ids"`
+	AllowedUserIDs []int `json:"allowed_user_ids"`
 }
 
-func (c Configuration) isAllowedUserID(id int64) bool {
+func (c Configuration) isAllowedUserID(id int) bool {
 	for _, userId := range c.AllowedUserIDs {
 		if userId == id {
 			return true
